Revert "Added support for project node selector"
