+++ pkg/cmd/admin/project/new_project.go
@@ -20,10 +20,9 @@ import (
 const NewProjectRecommendedName = "new-project"
 
 type NewProjectOptions struct {
-	ProjectName  string
-	DisplayName  string
-	Description  string
-	NodeSelector string
+	ProjectName string
+	DisplayName string
+	Description string
 
 	Client client.Interface
 
@@ -60,7 +59,6 @@ func NewCmdNewProject(name, fullName string, f *clientcmd.Factory, out io.Writer
 	cmd.Flags().StringVar(&options.AdminUser, "admin", "", "project admin username")
 	cmd.Flags().StringVar(&options.DisplayName, "display-name", "", "project display name")
 	cmd.Flags().StringVar(&options.Description, "description", "", "project description")
-	cmd.Flags().StringVar(&options.NodeSelector, "node-selector", "", "Restrict pods onto nodes matching given label selector")
 
 	return cmd
 }
@@ -88,7 +86,6 @@ func (o *NewProjectOptions) Run() error {
 	project.Annotations = make(map[string]string)
 	project.Annotations["description"] = o.Description
 	project.Annotations["displayName"] = o.DisplayName
-	project.Annotations["nodeSelector"] = o.NodeSelector
 	project, err := o.Client.Projects().Create(project)
 	if err != nil {
 		return err
