+++ pkg/cmd/server/api/helpers.go
@@ -89,8 +89,6 @@ func GetMasterFileReferences(config *MasterConfig) []*string {
 
 	refs = append(refs, &config.PolicyConfig.BootstrapPolicyFile)
 
-	refs = append(refs, &config.ProjectNodeSelector)
-
 	return refs
 }
