+++ pkg/cmd/server/api/v1/types.go
@@ -81,8 +81,6 @@ type MasterConfig struct {
 
 	PolicyConfig PolicyConfig `json:"policyConfig"`
 
-	// ProjectNodeSelector holds default project node label selector
-	ProjectNodeSelector string `json:"projectNodeSelector,omitempty"`
 	// ProjectRequestConfig holds information about how to handle new project requests
 	ProjectRequestConfig ProjectRequestConfig `json:"projectRequestConfig"`
 }
