+++ pkg/cmd/server/kubernetes/master.go
@@ -26,7 +26,6 @@ import (
 	"github.com/GoogleCloudPlatform/kubernetes/plugin/pkg/scheduler/factory"
 
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/namespace"
-	originscheduler "github.com/openshift/origin/pkg/scheduler"
 )
 
 const (
@@ -172,5 +171,5 @@ func (c *MasterConfig) createSchedulerConfig() (*scheduler.Config, error) {
 	}
 
 	// if the config file isn't provided, use the default provider
-	return configFactory.CreateFromProvider(originscheduler.DefaultProvider)
+	return configFactory.CreateFromProvider(factory.DefaultProvider)
 }
