+++ pkg/cmd/server/start/plugins.go
@@ -8,8 +8,5 @@ import (
 	_ "github.com/GoogleCloudPlatform/kubernetes/plugin/pkg/admission/namespace/exists"
 	_ "github.com/GoogleCloudPlatform/kubernetes/plugin/pkg/admission/namespace/lifecycle"
 	_ "github.com/GoogleCloudPlatform/kubernetes/plugin/pkg/admission/resourcequota"
-	_ "github.com/openshift/origin/pkg/project/admission/lifecycle"
-
-	// Scheduler plugins used by OpenShift
-	_ "github.com/openshift/origin/pkg/scheduler/algorithmprovider/defaults"
+	_ "github.com/openshift/origin/pkg/project/admission"
 )
