+++ pkg/project/admission/admission.go
@@ -25,20 +25,25 @@ import (
 	apierrors "github.com/GoogleCloudPlatform/kubernetes/pkg/api/errors"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/api/meta"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/client"
-
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/client/cache"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/fields"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/labels"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/runtime"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/watch"
 	"github.com/openshift/origin/pkg/api/latest"
-	"github.com/openshift/origin/pkg/project/cache"
 )
 
 // TODO: modify the upstream plug-in so this can be collapsed
 // need ability to specify a RESTMapper on upstream version
 func init() {
 	admission.RegisterPlugin("OriginNamespaceLifecycle", func(client client.Interface, config io.Reader) (admission.Interface, error) {
-		return NewLifecycle()
+		return NewLifecycle(client), nil
 	})
 }
 
 type lifecycle struct {
+	client client.Interface
+	store  cache.Store
 }
 
 // Admit enforces that a namespace must exist in order to associate content with it.
@@ -61,19 +66,35 @@ func (e *lifecycle) Admit(a admission.Attributes) (err error) {
 		return nil
 	}
 
+	// check for namespace in the cache
+	namespaceObj, exists, err := e.store.Get(&kapi.Namespace{
+		ObjectMeta: kapi.ObjectMeta{
+			Name:      a.GetNamespace(),
+			Namespace: "",
+		},
+		Status: kapi.NamespaceStatus{},
+	})
+
+	if err != nil {
+		return err
+	}
+
 	name := "Unknown"
 	obj := a.GetObject()
 	if obj != nil {
 		name, _ = meta.NewAccessor().Name(obj)
 	}
 
-	projects, err := cache.GetProjectCache()
-	if err != nil {
-		return err
-	}
-	namespace, err := projects.GetNamespaceObject(a.GetNamespace())
-	if err != nil {
-		return apierrors.NewForbidden(kind, name, err)
+	var namespace *kapi.Namespace
+	if exists {
+		namespace = namespaceObj.(*kapi.Namespace)
+	} else {
+		// Our watch maybe latent, so we make a best effort to get the object, and only fail if not found
+		namespace, err = e.client.Namespaces().Get(a.GetNamespace())
+		// the namespace does not exist, so prevent create and update in that namespace
+		if err != nil {
+			return apierrors.NewForbidden(kind, name, fmt.Errorf("Namespace %s does not exist", a.GetNamespace()))
+		}
 	}
 
 	if a.GetOperation() != "CREATE" {
@@ -87,6 +108,24 @@ func (e *lifecycle) Admit(a admission.Attributes) (err error) {
 	return apierrors.NewForbidden(kind, name, fmt.Errorf("Namespace %s is terminating", a.GetNamespace()))
 }
 
-func NewLifecycle() (admission.Interface, error) {
-	return &lifecycle{}, nil
+func NewLifecycle(c client.Interface) admission.Interface {
+	store := cache.NewStore(cache.MetaNamespaceKeyFunc)
+	reflector := cache.NewReflector(
+		&cache.ListWatch{
+			ListFunc: func() (runtime.Object, error) {
+				return c.Namespaces().List(labels.Everything(), fields.Everything())
+			},
+			WatchFunc: func(resourceVersion string) (watch.Interface, error) {
+				return c.Namespaces().Watch(labels.Everything(), fields.Everything(), resourceVersion)
+			},
+		},
+		&kapi.Namespace{},
+		store,
+		0,
+	)
+	reflector.Run()
+	return &lifecycle{
+		client: c,
+		store:  store,
+	}
 }
