+++ pkg/project/api/validation/validation.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 
 	kvalidation "github.com/GoogleCloudPlatform/kubernetes/pkg/api/validation"
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/labels"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util/fielderrors"
 	"github.com/openshift/origin/pkg/project/api"
@@ -24,7 +23,6 @@ func ValidateProject(project *api.Project) fielderrors.ValidationErrorList {
 	if !validateNoNewLineOrTab(project.Annotations["displayName"]) {
 		result = append(result, fielderrors.NewFieldInvalid("displayName", project.Annotations["displayName"], "may not contain a new line or tab"))
 	}
-	result = append(result, validateNodeSelector(project)...)
 	return result
 }
 
@@ -37,7 +35,6 @@ func validateNoNewLineOrTab(s string) bool {
 func ValidateProjectUpdate(newProject *api.Project, oldProject *api.Project) fielderrors.ValidationErrorList {
 	allErrs := fielderrors.ValidationErrorList{}
 	allErrs = append(allErrs, kvalidation.ValidateObjectMetaUpdate(&oldProject.ObjectMeta, &newProject.ObjectMeta).Prefix("metadata")...)
-	allErrs = append(allErrs, validateNodeSelector(newProject)...)
 	newProject.Spec.Finalizers = oldProject.Spec.Finalizers
 	newProject.Status = oldProject.Status
 	return allErrs
@@ -48,17 +45,4 @@ func ValidateProjectRequest(request *api.ProjectRequest) fielderrors.ValidationE
 	project.ObjectMeta = request.ObjectMeta
 
 	return ValidateProject(project)
-}
-
-func validateNodeSelector(p *api.Project) fielderrors.ValidationErrorList {
-	allErrs := fielderrors.ValidationErrorList{}
-
-	if len(p.Annotations) > 0 {
-		if selector, ok := p.Annotations["nodeSelector"]; ok {
-			if _, err := labels.Parse(selector); err != nil {
-				allErrs = append(allErrs, fielderrors.NewFieldInvalid("nodeSelector", p.Annotations["nodeSelector"], "must be a valid label selector"))
-			}
-		}
-	}
-	return allErrs
 }
