--- pkg/scheduler/algorithmprovider/plugins.go
@@ -1,6 +1,0 @@
-// This package is used to register algorithm provider plugins.
-package algorithmprovider
-
-import (
-	_ "github.com/openshift/origin/pkg/scheduler/algorithmprovider/defaults"
-)
