+++ test/integration/project_test.go
@@ -131,8 +131,7 @@ func TestProjectIsNamespace(t *testing.T) {
 		ObjectMeta: kapi.ObjectMeta{
 			Name: "new-project",
 			Annotations: map[string]string{
-				"displayName":  "Hello World",
-				"nodeSelector": "env=test",
+				"displayName": "Hello World",
 			},
 		},
 	}
@@ -152,9 +151,7 @@ func TestProjectIsNamespace(t *testing.T) {
 	if project.Annotations["displayName"] != namespace.Annotations["displayName"] {
 		t.Fatalf("Project display name did not match namespace annotation, project %v, namespace %v", project.Annotations["displayName"], namespace.Annotations["displayName"])
 	}
-	if project.Annotations["nodeSelector"] != namespace.Annotations["nodeSelector"] {
-		t.Fatalf("Project node selector did not match namespace node selector, project %v, namespace %v", project.Annotations["nodeSelector"], namespace.Annotations["displayname"])
-	}
+
 }
 
 // TestProjectMustExist verifies that content cannot be added in a project that does not exist
@@ -213,4 +210,5 @@ func TestProjectMustExist(t *testing.T) {
 	if err == nil {
 		t.Errorf("Expected an error on creation of a Origin resource because namespace does not exist")
 	}
+
 }
